refac: move domain to internal
