+++ internal/domain/errors.go
